refactor: rename MinIOStorage → S3CompatibleStorage
The implementation has always been S3-compatible (MinIO and AWS S3 share
the same wire protocol). The original name was a historical artifact from
when MinIO was the only deployment target.

Changes:
- Rename omnibenchmark/remote/MinIOStorage.py → S3Storage.py
- Rename class MinIOStorage → S3CompatibleStorage
- Add MinIOStorage alias at module level for backward compatibility
  (existing code importing MinIOStorage will continue to work)
- Update storage.py, cli/remote.py and test imports accordingly
